refactor(use custom attributes query): rename file & simply useMemo usage
